Update venture_funding_with_deep_learning.ipynb
Re-Defined variables to include A1 in description. Created the First hidden layer, output layer and checked the structure of the 1st model. Compiled the sequential model. Fit the model using 50 epochs and the new training data. Displayed the accuracy of the Orriginal model and the alternate model 1 and compared the results